+++ src/extension.ts
@@ -106,6 +106,9 @@ export function activate(context: vscode.ExtensionContext): void {
     vscode.commands.registerCommand("rojoExplorer.createLocalizationTable", (node?: ExplorerNode) =>
       createResource(provider, fileSystem, node, "LocalizationTable"),
     ),
+    vscode.commands.registerCommand("rojoExplorer.createJsonModule", (node?: ExplorerNode) =>
+      createResource(provider, fileSystem, node, "JSONModule"),
+    ),
     vscode.commands.registerCommand("rojoExplorer.renameResource", (node?: ExplorerNode) =>
       renameResource(provider, fileSystem, node),
     ),
@@ -764,6 +767,8 @@ function localizeResourceKind(kind: CreatableResourceKind): string {
       return vscode.l10n.t("StringValue");
     case "LocalizationTable":
       return vscode.l10n.t("LocalizationTable");
+    case "JSONModule":
+      return vscode.l10n.t("JSON Module");
   }
 }